Safe delete for python folder on Windows

+++ recipes/capypdf/all/conanfile.py
@@ -77,7 +77,8 @@ class PackageConan(ConanFile):
         rmdir(self, os.path.join(self.package_folder, "lib", "pkgconfig"))
         folders = os.listdir(os.path.join(self.package_folder, "lib",))
         python_folder = [f for f in folders if f.startswith("python") and os.path.isdir(os.path.join(self.package_folder, "lib", f))]
-        rmdir(self, os.path.join(self.package_folder, "lib", python_folder[0]))
+        if python_folder:
+            rmdir(self, os.path.join(self.package_folder, "lib", python_folder[0]))
         rm(self, "*.pdb", self.package_folder, recursive=True)
 
         fix_apple_shared_install_name(self)
